Enable lightning reports in configurations and add command handling in Static.js. Also added a few additional commands for testing.
