+++ configurations.json
@@ -87,7 +87,7 @@
                 "endpoint": "https://placefilenation.com/Placefiles/mcd.php"
             },
             "lightning_reports": {
-                "enabled": false,
+                "enabled": true,
                 "cache_time": 119,
                 "endpoint": "https://placefilenation.com/Placefiles/20lightning.php"
             }
